Added scope benchmarks
